RT-108 Move request_body_schema
