+++ mhs/outbound/generate_openapi.py
@@ -5,7 +5,8 @@ import apispec.ext.marshmallow
 import apispec_webframeworks.tornado
 from mhs_common.request import healthcheck_handler
 
-from outbound.request.synchronous import request_body_schema, handler
+from outbound.request import request_body_schema
+from outbound.request.synchronous import handler
 
 
 def create_spec() -> apispec.APISpec:
